docx-toolbox/requirements.txt
# Core dependencies
python-docx>=1.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
matplotlib>=3.8.0
reportlab>=4.0.0

# Optional GUI dependencies (all included for full local environment)
PySide6>=6.6.0
flet>=0.81.0
pywebview>=5.0.0

# Development and packaging dependencies
pytest>=8.0.0
pytest-cov>=5.0.0
pyinstaller>=6.0.0
setuptools>=68.0
wheel
